api/requirements.txt
fastapi
uvicorn
python-dotenv
spotipy
requests
openai
python-multipart
